move prom generation into it's own file
