+++ source/server/http/stats_handler.h
@@ -66,97 +66,5 @@ private:
                                  bool pretty_print = false);
 };
 
-/**
- * Formatter for metric/labels exported to Prometheus.
- *
- * See: https://prometheus.io/docs/concepts/data_model
- */
-class PrometheusStatsFormatter {
-public:
-  /**
-   * Extracts counters and gauges and relevant tags, appending them to
-   * the response buffer after sanitizing the metric / label names.
-   * @return uint64_t total number of metric types inserted in response.
-   */
-  static uint64_t statsAsPrometheus(const std::vector<Stats::CounterSharedPtr>& counters,
-                                    const std::vector<Stats::GaugeSharedPtr>& gauges,
-                                    const std::vector<Stats::ParentHistogramSharedPtr>& histograms,
-                                    Buffer::Instance& response, const bool used_only,
-                                    const absl::optional<std::regex>& regex);
-  /**
-   * Format the given tags, returning a string as a comma-separated list
-   * of <tag_name>="<tag_value>" pairs.
-   */
-  static std::string formattedTags(const std::vector<Stats::Tag>& tags);
-  /**
-   * Format the given metric name, prefixed with "envoy_".
-   */
-  static std::string metricName(const std::string& extracted_name);
-
-private:
-  /**
-   * Take a string and sanitize it according to Prometheus conventions.
-   */
-  static std::string sanitizeName(const std::string& name);
-
-  /*
-   * Determine whether a metric has never been emitted and choose to
-   * not show it if we only wanted used metrics.
-   */
-  template <class StatType>
-  static bool shouldShowMetric(const StatType& metric, const bool used_only,
-                               const absl::optional<std::regex>& regex) {
-    return ((!used_only || metric.used()) &&
-            (!regex.has_value() || std::regex_search(metric.name(), regex.value())));
-  }
-
-  /**
-   * Processes a stat type (counter, gauge, histogram) by generating all output lines, sorting
-   * them by tag-extracted metric name, and then outputting them in the correct sorted order into
-   * response.
-   *
-   * @param response The buffer to put the output into.
-   * @param used_only Whether to only output stats that are used.
-   * @param regex A filter on which stats to output.
-   * @param metrics The metrics to output stats for. This must contain all stats of the given type
-   *        to be included in the same output.
-   * @param generate_output A function which returns the output text for this metric.
-   * @param type The name of the prometheus metric type for used in TYPE annotations.
-   */
-  template <class StatType>
-  static uint64_t outputStatType(
-      Buffer::Instance& response, const bool used_only, const absl::optional<std::regex>& regex,
-      const std::vector<Stats::RefcountPtr<StatType>>& metrics,
-      const std::function<std::string(
-          const StatType& metric, const std::string& prefixed_tag_extracted_name)>& generate_output,
-      absl::string_view type);
-
-  /*
-   * Return the prometheus output for a numeric Stat (Counter or Gauge).
-   */
-  template <class StatType>
-  static std::string generateNumericOutput(const StatType& metric,
-                                           const std::string& prefixed_tag_extracted_name);
-
-  /*
-   * Returns the prometheus output for a histogram. The output is a multi-line string (with embedded
-   * newlines) that contains all the individual bucket counts and sum/count for a single histogram
-   * (metric_name plus all tags).
-   */
-  static std::string generateHistogramOutput(const Stats::ParentHistogram& histogram,
-                                             const std::string& prefixed_tag_extracted_name);
-
-  /*
-   * Comparator for Stats::Metric that does not require a string representation
-   * to make the comparison, for memory efficiency.
-   */
-  struct MetricLessThan {
-    bool operator()(const Stats::Metric* a, const Stats::Metric* b) const {
-      ASSERT(&a->constSymbolTable() == &b->constSymbolTable());
-      return a->constSymbolTable().lessThan(a->statName(), b->statName());
-    }
-  };
-};
-
 } // namespace Server
 } // namespace Envoy
