+++ tools/code_format/check_format.py
@@ -79,7 +79,8 @@ STD_REGEX_WHITELIST = (
     "./source/extensions/filters/http/squash/squash_filter.h",
     "./source/extensions/filters/http/squash/squash_filter.cc", "./source/server/http/utils.h",
     "./source/server/http/utils.cc", "./source/server/http/stats_handler.h",
-    "./source/server/http/stats_handler.cc", "./tools/clang_tools/api_booster/main.cc",
+    "./source/server/http/stats_handler.cc", "./source/server/http/prometheus_stats.h",
+    "./source/server/http/prometheus_stats.cc", "./tools/clang_tools/api_booster/main.cc",
     "./tools/clang_tools/api_booster/proto_cxx_utils.cc", "./source/common/common/version.cc")
 
 # Only one C++ file should instantiate grpc_init
